refactor: ad InputView and OutputView classes
- add two classes to separate business logic from UI logic
- modidy existing classes to call function in other classes from main

These changes adhere to the programming requirements 3
